move canvas token code to interfaces
